Create README.md
The basic approach to this lab was to first develop and debug the system using the simulator and then interface with actual lights and switches on a physical TM4C123. The simulator requires a different amount of actual time as compared to simulated time. On the other hand, the correct simulation time is maintained in the SysTick timer, which is decremented every cycle. The simulator speed depends on the amount of information it needs to update into the windows and the speed of the personal computer. Because we do not want to wait the minutes required for an actual intersection, the cars in this traffic intersection travel much faster than real cars.